refactor: Organize servicios commons de activity en subcarpetas
