+++ src/main/java/trinity/play2learn/backend/activity/activity/services/commons/count/ActivityCountByStudent.java
@@ -1,4 +1,4 @@
-package trinity.play2learn.backend.activity.activity.services.commons;
+package trinity.play2learn.backend.activity.activity.services.commons.count;
 
 import java.util.List;
