+++ src/main/java/trinity/play2learn/backend/activity/activity/services/commons/count/ActivityCountCompletedService.java
@@ -1,4 +1,4 @@
-package trinity.play2learn.backend.activity.activity.services.commons;
+package trinity.play2learn.backend.activity.activity.services.commons.count;
 
 import org.springframework.stereotype.Service;
